pyproject.toml
[project]
name = "ddgo-backup"
version = "1.0.0"
description = "Herramienta para exportar contraseñas de DuckDuckGo Android usando el Recovery Code"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Homero Thompson del Lago del Terror" }
]
keywords = ["duckduckgo", "password", "backup", "export", "sync"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Utilities",
]
dependencies = [
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "pynacl>=1.6.2",
]

[project.scripts]
ddgo-backup = "ddgo_backup.main:main"

[project.urls]
Homepage = "https://github.com/user/ddgo-backup"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ddgo_backup"]

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_ignores = true
